services/harvester/requirements.txt
# Dapr Agents Framework - Local development version with latest MCP transport
-e /workspaces/compliance-sentinel/dapr-agents

# Core Dapr and FastAPI dependencies
dapr>=1.14.0
dapr-ext-fastapi>=1.14.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# HTTP and web scraping
httpx>=0.25.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
requests>=2.31.0

# NLP and text processing
nltk>=3.8.0
spacy>=3.7.0
textblob>=0.17.0
transformers>=4.35.0
torch>=2.1.0

# Data processing and utilities
pandas>=2.1.0
numpy>=1.24.0
python-dateutil>=2.8.0
python-multipart>=0.0.6

# Async and concurrency
asyncio-mqtt>=0.13.0
aiofiles>=23.2.0

# Configuration and environment
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Logging and monitoring
structlog>=23.2.0
prometheus-client>=0.19.0

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Security
cryptography>=41.0.0
